Class for a list of MaestroPipelines A MaestroPipelineList is created when there are multiple maestro pipelines in a single script

Class for a list of MaestroPipelines A MaestroPipelineList is created when there are multiple maestro pipelines in a single script

Class for a list of MaestroPipelines A MaestroPipelineList is created when there are multiple maestro pipelines in a single script

Public fields

MaestroPipelines

list of pipelines

n_pipelines

number of pipelines in the list

Methods

Public methods

MaestroPipelineList$new()

MaestroPipelineList$print()

MaestroPipelineList$add_pipelines()

MaestroPipelineList$get_pipe_names()

MaestroPipelineList$get_schedule()

MaestroPipelineList$get_timely_pipelines()

MaestroPipelineList$check_timeliness()

MaestroPipelineList$get_status()

MaestroPipelineList$get_errors()

MaestroPipelineList$get_warnings()

MaestroPipelineList$get_messages()

MaestroPipelineList$get_artifacts()

MaestroPipelineList$run()

MaestroPipelineList$clone()

Method new()

Create a MaestroPipelineList object

Usage

MaestroPipelineList$new(MaestroPipelines = list())

Arguments

MaestroPipelines

list of MaestroPipelines

Returns

MaestroPipelineList

Method print()

Print the MaestroPipelineList

Usage

MaestroPipelineList$print()

Returns

print

Method add_pipelines()

Add pipelines to the list

Usage

MaestroPipelineList$add_pipelines(MaestroPipelines = NULL)

Arguments

MaestroPipelines

list of MaestroPipelines

Returns

invisible

Method get_pipe_names()

Get names of the pipelines in the list

Usage

MaestroPipelineList$get_pipe_names()

Returns

character

Method get_schedule()

Get the schedule as a data.frame

Usage

MaestroPipelineList$get_schedule()

Returns

data.frame

Method get_timely_pipelines()

Get a new MaestroPipelineList containing only those pipelines scheduled to run

Usage

MaestroPipelineList$get_timely_pipelines(...)

Arguments

...

arguments passed to self$check_timeliness

Returns

MaestroPipelineList

Method check_timeliness()

Check whether pipelines in the list are scheduled to run based on orchestrator frequency and current time

Usage

MaestroPipelineList$check_timeliness(...)

Arguments

...

arguments passed to self$check_timeliness

Returns

logical

Method get_status()

Get status of the pipelines as a data.frame

Usage

MaestroPipelineList$get_status()

Returns

data.frame

Method get_errors()

Get list of errors from the pipelines

Usage

MaestroPipelineList$get_errors()

Returns

list

Method get_warnings()

Get list of warnings from the pipelines

Usage

MaestroPipelineList$get_warnings()

Returns

list

Method get_messages()

Get list of messages from the pipelines

Usage

MaestroPipelineList$get_messages()

Returns

list

Method get_artifacts()

Get artifacts (return values) from the pipelines

Usage

MaestroPipelineList$get_artifacts()

Returns

list

Method run()

Runs all the pipelines in the list

Usage

MaestroPipelineList$run(..., cores = 1L)

Arguments

...

arguments passed to MaestroPipeline$run

cores

if using multicore number of cores to run in (uses furrr)

Returns

invisible

Method clone()

The objects of this class are cloneable with this method.

Usage

MaestroPipelineList$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.