Retrieve latest maestro build errors

Gets the latest schedule build errors following use of build_schedule(). If the build succeeded or build_schedule() has not been run it will be NULL.

Usage

last_build_errors()

Value

error messages

Examples

pipeline_dir <- tempdir()
create_pipeline("my_new_pipeline", pipeline_dir, open = FALSE, overwrite = TRUE)
#> ! Overwriting existing pipeline at /tmp/RtmpMs2pZh/my_new_pipeline.R.
#> ✔ Created pipeline at /tmp/RtmpMs2pZh/my_new_pipeline.R
build_schedule(pipeline_dir = pipeline_dir)
#> ℹ 3 scripts successfully parsed
#> Warning: 1 script failed to parse:
#> ! Could not build schedule entry with error: No directory called ./pipelines
#> ℹ See full error output with `last_build_errors()`
#> # A tibble: 3 × 8
#>   script_path            pipe_name frequency start_time          skip  log_level
#>   <chr>                  <chr>     <chr>     <dttm>              <lgl> <chr>    
#> 1 /tmp/RtmpMs2pZh/extra… extract_… 1 hour    2024-08-08 00:00:00 FALSE INFO     
#> 2 /tmp/RtmpMs2pZh/my_ne… my_new_p… 1 day     2024-08-08 00:00:00 FALSE INFO     
#> 3 /tmp/RtmpMs2pZh/new_j… new_job   20 minut… 2024-06-21 12:20:00 FALSE ERROR    
#> # ℹ 2 more variables: frequency_n <int>, frequency_unit <chr>
last_build_errors()
#> $orchestrator.R
#> <error/rlang_error>
#> Error:
#> ! Could not build schedule entry with error: No directory called
#>   ./pipelines
#> ℹ Pipeline scripts should not typically include code that's run outside of a
#>   function. Be sure to wrap code in a function block.
#> ---
#> Backtrace:
#>      ▆
#>   1. └─pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
#>   2.   └─pkgdown::build_site(...)
#>   3.     └─pkgdown:::build_site_local(...)
#>   4.       └─pkgdown::build_reference(...)
#>   5.         ├─pkgdown:::unwrap_purrr_error(...)
#>   6.         │ └─base::withCallingHandlers(...)
#>   7.         └─purrr::map(...)
#>   8.           └─purrr:::map_("list", .x, .f, ..., .progress = .progress)
#>   9.             ├─purrr:::with_indexed_errors(...)
#>  10.             │ └─base::withCallingHandlers(...)
#>  11.             ├─purrr:::call_with_cleanup(...)
#>  12.             └─pkgdown (local) .f(.x[[i]], ...)
#>  13.               ├─base::withCallingHandlers(...)
#>  14.               └─pkgdown:::data_reference_topic(...)
#>  15.                 └─pkgdown:::run_examples(...)
#>  16.                   └─pkgdown:::highlight_examples(code, topic, env = env)
#>  17.                     └─downlit::evaluate_and_highlight(...)
#>  18.                       └─evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#>  19.                         └─evaluate:::evaluate_call(...)
#>  20.                           ├─evaluate (local) timing_fn(...)
#>  21.                           ├─evaluate (local) handle(...)
#>  22.                           │ └─base::try(f, silent = TRUE)
#>  23.                           │   └─base::tryCatch(...)
#>  24.                           │     └─base (local) tryCatchList(expr, classes, parentenv, handlers)
#>  25.                           │       └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  26.                           │         └─base (local) doTryCatch(return(expr), name, parentenv, handler)
#>  27.                           ├─base::withCallingHandlers(...)
#>  28.                           ├─base::withVisible(...)
#>  29.                           └─evaluate:::eval_with_user_handlers(expr, envir, enclos, user_handlers)
#>  30.                             └─base::eval(expr, envir, enclos)
#>  31.                               └─base::eval(expr, envir, enclos)
#>  32.                                 └─maestro::build_schedule(pipeline_dir = pipeline_dir)
#>  33.                                   ├─stats::setNames(...)
#>  34.                                   └─purrr::map(pipelines, purrr::safely(build_schedule_entry))
#>  35.                                     └─purrr:::map_("list", .x, .f, ..., .progress = .progress)
#>  36.                                       ├─purrr:::with_indexed_errors(...)
#>  37.                                       │ └─base::withCallingHandlers(...)
#>  38.                                       ├─purrr:::call_with_cleanup(...)
#>  39.                                       └─purrr (local) .f(.x[[i]], ...)
#>  40.                                         ├─purrr:::capture_error(.f(...), otherwise, quiet)
#>  41.                                         │ └─base::tryCatch(...)
#>  42.                                         │   └─base (local) tryCatchList(expr, classes, parentenv, handlers)
#>  43.                                         │     └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  44.                                         │       └─base (local) doTryCatch(return(expr), name, parentenv, handler)
#>  45.                                         └─maestro (local) .f(...)
#>  46.                                           ├─base::suppressMessages(...)
#>  47.                                           │ └─base::withCallingHandlers(...)
#>  48.                                           └─base::tryCatch(...)
#>  49.                                             └─base (local) tryCatchList(expr, classes, parentenv, handlers)
#>  50.                                               ├─base (local) tryCatchOne(...)
#>  51.                                               │ └─base (local) doTryCatch(return(expr), name, parentenv, handler)
#>  52.                                               └─base (local) tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>  53.                                                 └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  54.                                                   └─value[[3L]](cond)
#>