Run a schedule

Given a schedule in a maestro project, runs the pipelines that are scheduled to execute based on the current time.

Usage

run_schedule(
  schedule,
  orch_frequency = "1 day",
  check_datetime = lubridate::now(tzone = "UTC"),
  resources = list(),
  run_all = FALSE,
  n_show_next = 5,
  cores = 1,
  logging = FALSE,
  log_file = "./maestro.log",
  log_file_max_bytes = 1e+06,
  quiet = FALSE
)

Arguments

schedule

a table of scheduled pipelines generated from build_schedule()

orch_frequency

of the orchestrator, a single string formatted like "1 day" or "2 weeks"

check_datetime

datetime to be used to check if the pipeline should run

resources

named list of shared resources made available to pipelines as needed

run_all

run all pipelines regardless of the schedule (default is FALSE) - useful for testing. Does not apply to pipes with a maestroSkip tag.

n_show_next

show the next n scheduled pipes

cores

number of cpu cores to run if running in parallel. If > 1, furrr is used and a multisession plan must be executed in the orchestrator (see details)

logging

whether or not to write the logs to a file (default = FALSE)

log_file

path to the log file (ignored if logging == FALSE)

log_file_max_bytes

numeric specifying the maximum number of bytes allowed in the log file before purging the log (within a margin of error)

quiet

silence metrics to the console (default = FALSE)

Value

data.frame of pipeline statuses

Details

Pipeline schedule logic

The function run_schedule() examines each pipeline in the schedule table and determines whether it is scheduled to run at the current time using some simple time arithmetic. We assume run_schedule(schedule, check_datetime = Sys.time()), but this need not be the case.

Pipelines with arguments (resources)

If a pipeline takes an argument that doesn't include a default value, these can be supplied in the orchestrator via run_schedule(resources = list(arg1 = val)). The name of the argument used by the pipeline must match the name of the argument in the list. Currently, each named resource must refer to a single object. In other words, you can't have two pipes using the same argument but requiring different values.

Running in parallel

Pipelines can be run in parallel using the cores argument. First, you must run future::plan(future::multisession) in the orchestrator. Then, supply the desired number of cores to the cores argument. Note that console output appears different in multicore mode.

Examples

pipeline_dir <- tempdir()
create_pipeline("my_new_pipeline", pipeline_dir, open = FALSE)
#> File /tmp/RtmpjWzInB/my_new_pipeline.R already exists. Overwrite? [Y/n]: 
schedule <- build_schedule(pipeline_dir = pipeline_dir)
#> ℹ 1 script successfully parsed
#> Warning: 1 script failed to parse:
#> ! Could not build schedule entry with error: No directory called ./pipelines
#> ℹ See full error output with `last_build_errors()`
# Runs the schedule every 1 day
run_schedule(
  schedule,
  orch_frequency = "1 day",
  quiet = TRUE
)
#> # A tibble: 1 × 10
#>   pipe_name  script_path invoked success pipeline_started    pipeline_ended     
#>   <chr>      <chr>       <lgl>   <lgl>   <dttm>              <dttm>             
#> 1 my_new_pi… /tmp/Rtmpj… TRUE    TRUE    2024-05-31 22:47:15 2024-05-31 22:47:15
#> # ℹ 4 more variables: errors <int>, warnings <int>, messages <int>,
#> #   next_run <dttm>

# Runs the schedule every 15 minutes
run_schedule(
  schedule,
  orch_frequency = "15 minutes",
  quiet = TRUE
)
#> # A tibble: 1 × 10
#>   pipe_name   script_path invoked success pipeline_started pipeline_ended errors
#>   <chr>       <chr>       <lgl>   <lgl>   <lgl>            <lgl>           <int>
#> 1 my_new_pip… /tmp/Rtmpj… FALSE   NA      NA               NA                  0
#> # ℹ 3 more variables: warnings <int>, messages <int>, next_run <dttm>